+++ tests/test_hookee_manager_edges.py
@@ -1,20 +1,20 @@
 import time
 
 import requests
-from hookee import conf
+from hookee.conf import Config
 
-from hookee.hookeemanager import HookeeManager
+from hookee import HookeeManager
 from tests.testcase import HookeeTestCase
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "1.1.0"
+__version__ = "1.2.0"
 
 
 class TestHookeeManagerEdges(HookeeTestCase):
     def test_hookee_manager(self):
         # GIVEN
-        hookee_manager = HookeeManager(self.ctx)
+        hookee_manager = HookeeManager()
         hookee_manager._init_server_and_tunnel()
 
         self.assertIsNotNone(hookee_manager.server._thread)
@@ -33,7 +33,8 @@ class TestHookeeManagerEdges(HookeeTestCase):
     def test_custom_response(self):
         # GIVEN
         response_body = "<Response>Ok</Response>"
-        hookee_manager = HookeeManager(conf.Context({"response": response_body}))
+        config = Config(response=response_body)
+        hookee_manager = HookeeManager(config=config)
         hookee_manager._init_server_and_tunnel()
 
         webhook_url = "{}/webhook".format(hookee_manager.tunnel.public_url)