+++ tests/test_util.py
@@ -1,4 +1,4 @@
-from hookee.hookeemanager import HookeeManager
+from hookee import HookeeManager
 from hookee.pluginmanager import PluginManager
 
 from hookee import util
@@ -6,14 +6,14 @@ from tests.testcase import HookeeTestCase
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "1.1.0"
+__version__ = "1.2.0"
 
 
 class TestUtil(HookeeTestCase):
     def setUp(self):
         super(TestUtil, self).setUp()
 
-        self.hookee_manager = HookeeManager(self.ctx)
+        self.hookee_manager = HookeeManager()
         self.plugin_manager = PluginManager(self.hookee_manager)
         self.plugin = self.plugin_manager.get_plugin("request_url_info")